Improve output of elem2str utility
Add proper support for some classes on Octave:
- string: Show the string encapsulated in quotation marks
- small numerics: use `mat2str` on the element

Since these outputs for string and numeric elements are more
informative than the output from `disp()` (it is useful to
indicate the class of these elements), these settings superceed
the `evalc(disp(elem))` method.

`evalc(disp(elem))` is still used on MATLAB for all other classes
(including cells).

On Octave, the fallback is still to indicate just the class of the
element, but now to also indicate the size if it is available.