Relaxed interface to not require id
I relaxed the API so an id is not required.  This is helpful when users aren't carefully tracking exception ID's, and is often the case because of poor enum support in Matlab.  In fact, many users don't use ID's at all, and requiring an ID for unit testing is likely to result in less unit testing.